requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dotenv-python==0.0.1
easyocr==1.4
Flask==2.0.1
idna==3.2
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.2
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.0
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
tifffile==2021.8.30
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
